requirements.txt
osmnx==0.7.3
numpy==1.14.2
tqdm==4.11.2
pyproj==1.9.5.1
Shapely==1.6.4.post1
recordclass==0.5
Cython==0.28
requests==2.18.4
scipy==0.19.1
geopandas==0.3.0
pandas==0.22.0
python-geohash==0.8.5
